create calculate/{applicationId} endpoint

+++ deal/src/main/resources/db/changelog/v-1.0/10-create-credit-table.sql
@@ -1,10 +1,10 @@
 create table credit (
     credit_id bigserial
-    , amount double precision not null
+    , amount numeric(100, 10) not null
     , term integer not null
-    , monthly_payment double precision not null
-    , rate double precision not null
-    , psk double precision not null
+    , monthly_payment numeric(100, 10) not null
+    , rate numeric(15, 10) not null
+    , psk numeric(100, 10) not null
     , payment_schedule jsonb not null
     , insurance_enable boolean not null
     , salary_client boolean not null
